refactor out partition id get/create
